Timeout
Remove DOTWEEN pro files

+++ Assets/Scripts/Commands/CostCommand.cs
@@ -1,8 +1,8 @@
 ﻿namespace Commands {
     public class CostCommand : Command {
-        public float Cost { get; }
+        public int Cost { get; }
 
-        public CostCommand(int lineNumber, string line, float cost) : base(lineNumber, line) {
+        public CostCommand(int lineNumber, string line, int cost) : base(lineNumber, line) {
             Cost = cost;
         }
     }
